refactor(api): núcleo de fetch compartilhado + cliente admin-api [US-14]
