+++ src/auth/auth-api.ts
@@ -1,41 +1,5 @@
-function readCookie(name: string): string | undefined {
-  return document.cookie.split('; ').find(c => c.startsWith(name + '='))?.split('=')[1]
-}
+import { makeApiClient, ensureCsrf } from './api-core.js'
 
-const MUTATING = new Set(['POST', 'PUT', 'PATCH', 'DELETE'])
-
-async function rawFetch(path: string, init: RequestInit = {}): Promise<Response> {
-  const method = (init.method ?? 'GET').toUpperCase()
-  const headers = new Headers(init.headers)
-  if (MUTATING.has(method)) {
-    const csrf = readCookie('csrf_token')
-    if (csrf) headers.set('X-CSRF-Token', decodeURIComponent(csrf))
-    if (init.body && !headers.has('Content-Type')) headers.set('Content-Type', 'application/json')
-  }
-  return fetch(`/api/auth${path}`, { ...init, headers, credentials: 'same-origin' })
-}
-
-let refreshing: Promise<boolean> | null = null
-async function tryRefresh(): Promise<boolean> {
-  if (!refreshing) {
-    refreshing = rawFetch('/refresh', { method: 'POST' })
-      .then(r => r.ok)
-      .catch(() => false)
-      .finally(() => { refreshing = null })
-  }
-  return refreshing
-}
-
-/** Garante que o cookie CSRF existe (chamar no boot). */
-export async function ensureCsrf(): Promise<void> {
-  if (!readCookie('csrf_token')) await rawFetch('/csrf')
-}
-
-/** Fetch com auto-refresh: em 401, tenta /refresh uma vez e repete. */
-export async function apiFetch(path: string, init: RequestInit = {}): Promise<Response> {
-  let res = await rawFetch(path, init)
-  if (res.status === 401 && path !== '/refresh' && path !== '/login') {
-    if (await tryRefresh()) res = await rawFetch(path, init)
-  }
-  return res
-}
+const client = makeApiClient('/api/auth')
+export const apiFetch = client.apiFetch
+export { ensureCsrf }
