+++ src/painel/admin-api.ts
@@ -1,0 +1,4 @@
+import { makeApiClient } from '@/auth/api-core'
+
+const client = makeApiClient('/api/admin')
+export const adminFetch = client.apiFetch
